+++ backend/src/main/java/hanglog/auth/JwtProvider.java
@@ -1,4 +1,4 @@
-package hanglog.member.domain.auth;
+package hanglog.auth;
 
 import static hanglog.global.exception.ExceptionCode.EXPIRED_PERIOD_ACCESS_TOKEN;
 import static hanglog.global.exception.ExceptionCode.EXPIRED_PERIOD_REFRESH_TOKEN;
@@ -7,7 +7,7 @@ import static hanglog.global.exception.ExceptionCode.INVALID_REFRESH_TOKEN;
 
 import hanglog.global.exception.ExpiredPeriodJwtException;
 import hanglog.global.exception.InvalidJwtException;
-import hanglog.member.domain.MemberTokens;
+import hanglog.auth.domain.MemberTokens;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.Header;
