+++ backend/src/main/java/hanglog/auth/domain/oauthprovider/OauthProviders.java
@@ -1,4 +1,4 @@
-package hanglog.member.domain.auth;
+package hanglog.auth.domain.oauthprovider;
 
 import static hanglog.global.exception.ExceptionCode.NOT_SUPPORTED_OAUTH_SERVICE;
