+++ backend/src/main/java/hanglog/auth/domain/oauthuserinfo/GoogleUserInfo.java
@@ -1,0 +1,33 @@
+package hanglog.auth.domain.oauthuserinfo;
+
+import static lombok.AccessLevel.PRIVATE;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = PRIVATE)
+@AllArgsConstructor
+public class GoogleUserInfo implements OauthUserInfo {
+
+    @JsonProperty("id")
+    private String socialLoginId;
+    @JsonProperty("name")
+    private String name;
+    @JsonProperty("picture")
+    private String picture;
+
+    public String getSocialLoginId() {
+        return socialLoginId;
+    }
+
+    @Override
+    public String getNickname() {
+        return name;
+    }
+
+    @Override
+    public String getImageUrl() {
+        return picture;
+    }
+}
