+++ backend/src/main/java/hanglog/auth/domain/oauthuserinfo/KakaoUserInfo.java
@@ -1,34 +1,41 @@
-package hanglog.member.domain.auth.google;
+package hanglog.auth.domain.oauthuserinfo;
 
 import static lombok.AccessLevel.PRIVATE;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import hanglog.member.domain.auth.UserInfo;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = PRIVATE)
 @AllArgsConstructor
-public class GoogleUserInfo implements UserInfo {
+public class KakaoUserInfo implements OauthUserInfo {
 
     @JsonProperty("id")
     private String socialLoginId;
-    @JsonProperty("name")
-    private String name;
-    @JsonProperty("picture")
-    private String picture;
+    @JsonProperty("properties")
+    private Properties properties;
 
+    @Override
     public String getSocialLoginId() {
         return socialLoginId;
     }
 
     @Override
     public String getNickname() {
-        return name;
+        return properties.name;
     }
 
     @Override
     public String getImageUrl() {
-        return picture;
+        return properties.image;
+    }
+
+    @NoArgsConstructor(access = PRIVATE)
+    private static class Properties {
+
+        @JsonProperty("nickname")
+        private String name;
+        @JsonProperty("profile_image")
+        private String image;
     }
 }
