--- backend/src/main/java/hanglog/member/domain/auth/kakao/KakaoUserInfo.java
@@ -1,42 +1,0 @@
-package hanglog.member.domain.auth.kakao;
-
-import static lombok.AccessLevel.PRIVATE;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import hanglog.member.domain.auth.UserInfo;
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = PRIVATE)
-@AllArgsConstructor
-public class KakaoUserInfo implements UserInfo {
-
-    @JsonProperty("id")
-    private String socialLoginId;
-    @JsonProperty("properties")
-    private Properties properties;
-
-    @Override
-    public String getSocialLoginId() {
-        return socialLoginId;
-    }
-
-    @Override
-    public String getNickname() {
-        return properties.name;
-    }
-
-    @Override
-    public String getImageUrl() {
-        return properties.image;
-    }
-
-    @NoArgsConstructor(access = PRIVATE)
-    private static class Properties {
-
-        @JsonProperty("nickname")
-        private String name;
-        @JsonProperty("profile_image")
-        private String image;
-    }
-}
